Move FXMLFileLoader to library (fixes #40)
- Moves FXMLFileLoader to library
- Removes FXMLFileLoader from driverstation and field
- Moves gui resources out of gui package(into root folder) in the field resource folder
