+++ sumobots/driverstation/src/test/java/org/inspirerobotics/sumobots/driverstation/GuiTest.java
@@ -1,7 +1,7 @@
 package org.inspirerobotics.sumobots.driverstation;
 
-import org.inspirerobotics.sumobots.driverstation.gui.FXMLFileLoader;
 import org.inspirerobotics.sumobots.driverstation.gui.MainScene;
+import org.inspirerobotics.sumobots.library.gui.FXMLFileLoader;
 import org.inspirerobotics.sumobots.library.gui.JavaFXInitTestRule;
 import org.junit.Rule;
 import org.junit.Test;
