+++ sumobots/field/src/main/resources/fxml/gameTab.fxml
@@ -12,7 +12,7 @@
 
 <fx:root minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0"
 	prefWidth="1280.0" type="javafx.scene.layout.AnchorPane" xmlns="http://javafx.com/javafx/8"
-	xmlns:fx="http://javafx.com/fxml/1" stylesheets="@../css/styles.css">
+	xmlns:fx="http://javafx.com/fxml/1" stylesheets="@../../../../../driverstation/src/main/resources/css/styles.css">
 	<children>
 		<BorderPane fx:id="mainBorderPane" prefHeight="200.0" prefWidth="200.0"
 					AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
@@ -26,7 +26,7 @@
 							AnchorPane.topAnchor="0.0" BorderPane.alignment="CENTER">
 							<bottom>
 								<TableView fx:id="connTable" prefHeight="200.0"
-									prefWidth="200.0" styleClass="stdTable" stylesheets="@../css/styles.css"
+									prefWidth="200.0" styleClass="stdTable" stylesheets="@../../../../../driverstation/src/main/resources/css/styles.css"
 									BorderPane.alignment="CENTER">
                                                            		    
 									<columnResizePolicy>
