+++ sumobots/field/src/test/java/org/inspirerobotics/sumobots/field/GuiTest.java
@@ -1,7 +1,7 @@
 package org.inspirerobotics.sumobots.field;
 
-import org.inspirerobotics.sumobots.field.gui.FXMLFileLoader;
 import org.inspirerobotics.sumobots.field.gui.RootGroup;
+import org.inspirerobotics.sumobots.library.gui.FXMLFileLoader;
 import org.inspirerobotics.sumobots.library.gui.JavaFXInitTestRule;
 import org.junit.Rule;
 import org.junit.Test;
