+++ sumobots/library/src/main/java/org/inspirerobotics/sumobots/library/gui/FXMLFileLoader.java
@@ -1,7 +1,7 @@
-package org.inspirerobotics.sumobots.field.gui;
+package org.inspirerobotics.sumobots.library.gui;
 
 import javafx.fxml.FXMLLoader;
-import org.inspirerobotics.sumobots.library.Resources;
+import org.inspirerobotics.sumobots.library.InternalLog;
 
 import java.io.IOException;
 import java.util.logging.Level;
@@ -9,7 +9,7 @@ import java.util.logging.Logger;
 
 public class FXMLFileLoader {
 
-	private static final Logger logger = Logger.getLogger(Resources.LOGGER_NAME);
+	private static final Logger logger = InternalLog.getLogger();
 
 	public static void load(String name, Object controller) {
 		load(name, controller, controller);
@@ -17,7 +17,7 @@ public class FXMLFileLoader {
 
 	public static void load(String name, Object controller, Object root) {
 		logger.fine("Loading " + name);
-		FXMLLoader fxmlLoader = new FXMLLoader(FXMLFileLoader.class.getClass().getResource("/gui/fxml/" + name));
+		FXMLLoader fxmlLoader = new FXMLLoader(FXMLFileLoader.class.getClass().getResource("/fxml/" + name));
 		fxmlLoader.setController(controller);
 		fxmlLoader.setRoot(root);
 
@@ -32,5 +32,4 @@ public class FXMLFileLoader {
 
 		logger.fine("Loaded " + name);
 	}
-
 }
